201025_01

+++ secureswitools/swianalysistoolkit/source/chainvalidityandinstallfilestatustools/dumpinstallfilestatustool/options.cpp
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2005-2009 Nokia Corporation and/or its subsidiary(-ies).
+* Copyright (c) 2005-2010 Nokia Corporation and/or its subsidiary(-ies).
 * All rights reserved.
 * This component and the accompanying materials are made available
 * under the terms of the License "Eclipse Public License v1.0"
@@ -263,7 +263,7 @@ const int Options::ConvertUserCapsToInt(vector <string>& aUserCapabilities)
 				{
 					if(capabilities.compare(listOfCapabilities[i])==0)
 					{
-					userCaps += pow(2,i);
+					userCaps += pow((double)2,i);
 					}
 				}
 			}
